Converted floats to ints
As discussed with Titi, this may prevent some issues with multiplayer
getting out of sync due to different floating point calculations.

+++ source/glest_game/types/unit_type.cpp
@@ -614,7 +614,7 @@ void UnitType::loaddd(int id,const string &dir, const TechTree *techTree,
 				}
 				
 				if(resourceNode->hasAttribute("amount-percentage")) {
-					resource.setAmountPercentage(resourceNode->getAttribute("amount-percentage")->getFloatValue());
+					resource.setAmountPercentage(resourceNode->getAttribute("amount-percentage")->getIntValue());
 				}
 				else {
 					resource.setAmountPercentage(0);
@@ -628,7 +628,7 @@ void UnitType::loaddd(int id,const string &dir, const TechTree *techTree,
 				}
 				
 				if(resourceNode->hasAttribute("loss-percentage")) {
-					resource.setLossPercentage(resourceNode->getAttribute("loss-percentage")->getFloatValue());
+					resource.setLossPercentage(resourceNode->getAttribute("loss-percentage")->getIntValue());
 				}
 				else {
 					resource.setLossPercentage(0);
